+++ vue/components/ui/molecules/index.js
@@ -8,6 +8,7 @@ import { ButtonIconAnimated } from "./button-icon-animated/button-icon-animated.
 import { ButtonIconDropdown } from "./button-icon-dropdown/button-icon-dropdown.vue";
 import { ButtonIconToggle } from "./button-icon-toggle/button-icon-toggle.vue";
 import { ButtonToggle } from "./button-toggle/button-toggle.vue";
+import { ButtonUpload } from "./button-upload/button-upload.vue";
 import { Calendar } from "./calendar/calendar.vue";
 import { Carousel } from "./carousel/carousel.vue";
 import { ChatMessage } from "./chat-message/chat-message.vue";
@@ -43,7 +44,6 @@ import { Tabs } from "./tabs/tabs.vue";
 import { Tooltip } from "./tooltip/tooltip.vue";
 import { TransferList } from "./transfer-list/transfer-list.vue";
 import { UploadArea } from "./upload-area/upload-area.vue";
-import { UploadButton } from "./upload-button/upload-button.vue";
 
 const install = Vue => {
     Vue.component("alert", Alert);
@@ -56,6 +56,7 @@ const install = Vue => {
     Vue.component("button-icon-dropdown", ButtonIconDropdown);
     Vue.component("button-icon-toggle", ButtonIconToggle);
     Vue.component("button-toggle", ButtonToggle);
+    Vue.component("button-upload", ButtonUpload);
     Vue.component("calendar", Calendar);
     Vue.component("carousel", Carousel);
     Vue.component("chat-message", ChatMessage);
@@ -91,7 +92,6 @@ const install = Vue => {
     Vue.component("tooltip", Tooltip);
     Vue.component("transfer-list", TransferList);
     Vue.component("upload-area", UploadArea);
-    Vue.component("upload-button", UploadButton);
 };
 
 export {
@@ -104,6 +104,7 @@ export {
     ButtonIconAnimated,
     ButtonIconDropdown,
     ButtonIconToggle,
+    ButtonUpload,
     Calendar,
     Carousel,
     ChatMessage,
@@ -138,8 +139,7 @@ export {
     Tabs,
     Tooltip,
     TransferList,
-    UploadArea,
-    UploadButton
+    UploadArea
 };
 
 export default install;